a program that finds and prints the first 98 Fibonacci numbers, starting with 1 and 2, followed by a new line

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -11,15 +11,15 @@
  */
 int main(void)
 {
-int a = 1, b = 2, next;
+unsigned long a = 1, b = 2, next;
 int count = 0;
 
-printf("%d, %d", a, b);
+printf("%lu, %lu", a, b);
 
 while (count < 96)  /* We've already printed the first two numbers */
 {
 next = a + b;
-printf(", %d", next);
+printf(", %lu", next);
 a = b;
 b = next;
 count++;
